Minor: renames [username] route to [handle]
